+++ app/Http/Controllers/admin/CityController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\pharmacylocation;
+namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
