+++ app/Http/Controllers/admin/DirectorateController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\pharmacylocation;
+namespace App\Http\Controllers\admin;
 use App\Models\Directorate;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
